+++ app/Http/Controllers/post/PostController.php
@@ -2,12 +2,13 @@
 
 namespace App\Http\Controllers\post;
 
+use App\Models\Tag;
+use App\Models\Category;
 use App\Models\post\Post;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
-use App\Models\Category;
-use App\Models\Tag;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Storage;
 
 class PostController extends Controller
@@ -46,6 +47,9 @@ class PostController extends Controller
      */
     public function store(Request $request)
     {
+        if(!Gate::allows('AdminAndEditor')){
+            return abort(403);
+        }
         $request->validate([
             'title' => 'required|max:255',
             'slug' => 'required|max:255',
@@ -184,6 +188,9 @@ class PostController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        if(!Gate::allows('AdminAndEditor')){
+            return abort(403);
+        }
         $request->validate([
             'title' => 'required|max:255',
             'slug' => 'required|max:255',
@@ -238,6 +245,9 @@ class PostController extends Controller
      */
     public function destroy(string $id)
     {
+        if(!Gate::allows('AdminAndEditor')){
+            return abort(403);
+        }
         $fileName = Post::find($id);
         if (Storage::disk('public')->exists($fileName->featured_image)) {
             Storage::disk('public')->delete($fileName->featured_image);
